Remove strapi files
