+++ src/api-requests/get-podcasts.ts
@@ -1,106 +1,87 @@
-import { ApolloQueryResult, gql } from '@apollo/client';
-
 import { UnderlyingClient } from '../types/client-factory';
 import { Podcast } from '../interfaces/podcast';
+import { GenericTagTypeResponse } from '../interfaces/tag-type-response';
+import { CSEdges, CSPodcast } from '../interfaces/contentstack';
+import { OtherTags } from '../interfaces/other-tags';
 
-const podcastFields = `
-    description
-    publishDate: originalPublishDate
-    title
-    slug
-    podcastFileUrl
-    thumbnailUrl
-    casted_slug
-    l1Product: l_1_product {
-        name
-        calculatedSlug: calculated_slug
-    }
-    l2Product: l_2_product {
-        name
-        calculatedSlug: calculated_slug
-    }
-    programmingLanguage: programming_language {
-        name
-        calculatedSlug: calculated_slug
-    }
-    technology: technology {
-        name
-        calculatedSlug: calculated_slug
-    }
-    otherTags: other_tags {
-        spokenLanguage: spoken_language {
-            name
-            calculatedSlug: calculated_slug
-        }
-        expertiseLevel: expertise_level {
-            name: level
-            calculatedSlug: calculated_slug
-        }
-        authorType: author_type {
-            name
-            calculatedSlug: calculated_slug
-        }
-    }
-    seo: SEO {
-        canonical_url
-        meta_description
-        og_description
-        og_image {
-            url
-        }
-        og_title
-        og_type
-        og_url
-        twitter_card
-        twitter_creator
-        twitter_description
-        twitter_image {
-            url
-        }
-        twitter_site
-        twitter_title
-    }
-`;
+import {
+    fetchAll,
+    extractFieldsFromNode,
+    extractFieldsFromNodes,
+    getSEO,
+    getOtherTags,
+} from './utils';
+import { allPodcastsQuery, podcastsBySlugQuery } from '../graphql/podcasts';
+
+const formatResponses = (csPodcasts: CSPodcast[]): Podcast[] => {
+    return csPodcasts.map(p => ({
+        // explicitly defined instead of using spread operator
+        // to ensure default value for undefined
+        description: p.description ? p.description : '',
+        publishDate: p.publishDate ? p.publishDate : '',
+        title: p.title ? p.title : '',
+        slug: p.slug ? p.slug : '',
+        l1Product: extractFieldsFromNode(p.l1Product as CSEdges<any>, [
+            'name',
+            'calculatedSlug',
+        ]) as GenericTagTypeResponse,
+        l2Product: extractFieldsFromNode(p.l2Product as CSEdges<any>, [
+            'name',
+            'calculatedSlug',
+        ]) as GenericTagTypeResponse,
+        programmingLanguage: extractFieldsFromNodes(
+            p.programmingLanguage as CSEdges<any>,
+            ['name', 'calculatedSlug']
+        ) as GenericTagTypeResponse[],
+        technology: extractFieldsFromNodes(p.technology as CSEdges<any>, [
+            'name',
+            'calculatedSlug',
+        ]) as GenericTagTypeResponse[],
+        otherTags: getOtherTags(p.otherTags) as OtherTags,
+        seo: getSEO(p.seo),
+        podcastFileUrl: p.podcastFileUrl ? p.podcastFileUrl : '',
+        // thumbnailUrl follow old format to return null value
+        // but cast as string to conform to the interface
+        thumbnailUrl: (p.thumbnailUrl ? p.thumbnailUrl : null) as string,
+        casted_slug: p.casted_slug ? p.casted_slug : '',
+    }));
+};
 
 /**
  * Returns a list of all articles.
  * @param client -  The Apollo REST client that will be used to make the request.
  */
 const getAllPodcastsFromAPI = async (
-    client: UnderlyingClient<'ApolloREST'>
+    client: UnderlyingClient<'ApolloGraphQL'>
 ): Promise<Podcast[]> => {
-    const query = gql`
-        query Podcasts {
-            podcasts @rest(type: "Podcast", path: "/podcasts?_limit=-1") {
-                ${podcastFields}
-            }
-        }
-    `;
+    const podcasts = (await fetchAll(
+        client,
+        allPodcastsQuery,
+        'podcasts'
+    )) as CSPodcast[];
+    const data = formatResponses(podcasts);
 
-    const { data }: ApolloQueryResult<{ podcasts: Podcast[] }> =
-        await client.query({ query });
-
-    console.log(JSON.stringify(data, null, 2));
-    return data.podcasts;
+    return data;
 };
 
 export const getPodcastBySlugFromAPI = async (
-    client: UnderlyingClient<'ApolloREST'>,
+    client: UnderlyingClient<'ApolloGraphQL'>,
     slug: string
 ): Promise<Podcast | null> => {
-    const query = gql`
-        query Podcasts {
-            podcasts @rest(type: "Podcast", path: "/podcasts?slug_eq=${slug}") {
-                ${podcastFields}
-            }
-        }
-    `;
-    const { data }: ApolloQueryResult<{ podcasts: Podcast[] }> =
-        await client.query({
-            query,
-        });
+    const variables = { slug };
+    const podcasts = (await fetchAll(
+        client,
+        podcastsBySlugQuery,
+        'podcasts',
+        variables
+    )) as CSPodcast[];
 
-    return data.podcasts.length > 0 ? data.podcasts[0] : null;
+    if (!podcasts) {
+        return null;
+    }
+
+    const data = formatResponses(podcasts)[0];
+    return data;
 };
 
 export default getAllPodcastsFromAPI;
